requirements.txt
rdkit-pypi
pandas
tqdm
